Introduce ImageGenerator abstraction for messenger mock flow
