+++ server/_core/imageService.ts
@@ -1,14 +1,36 @@
-import { getStyleById, type StyleId } from "./messengerStyles";
+import { STYLE_TO_DEMO_FILE, type Style } from "./messengerStyles";
 
-export function getMockGeneratedImage(styleId: StyleId, cursor = 0): { imageUrl: string; nextCursor: number } {
-  const style = getStyleById(styleId);
-  const index = cursor % style.mockResultUrls.length;
-
-  return {
-    imageUrl: style.mockResultUrls[index],
-    nextCursor: cursor + 1,
-  };
+export interface ImageGenerator {
+  generate(input: {
+    style: Style;
+    sourceImageUrl?: string;
+    psid: string;
+  }): Promise<{ imageUrl: string }>;
 }
 
-// Future integration point:
-// Replace getMockGeneratedImage with OpenAI/real generation logic and keep the webhook UX unchanged.
+function getBaseUrl(): string {
+  const configuredBaseUrl = process.env.APP_BASE_URL?.trim() ?? process.env.BASE_URL?.trim();
+
+  if (configuredBaseUrl && /^https?:\/\//.test(configuredBaseUrl)) {
+    return configuredBaseUrl;
+  }
+
+  return "http://localhost:3000";
+}
+
+function getMockImageForStyle(style: Style): string {
+  const filename = STYLE_TO_DEMO_FILE[style];
+  return `${getBaseUrl()}/demo/${filename}`;
+}
+
+class MockImageGenerator implements ImageGenerator {
+  async generate(input: { style: Style; sourceImageUrl?: string; psid: string }): Promise<{ imageUrl: string }> {
+    return {
+      imageUrl: getMockImageForStyle(input.style),
+    };
+  }
+}
+
+export function createImageGenerator(): ImageGenerator {
+  return new MockImageGenerator();
+}
